Clean up bergamot translator integration somewhat so that the dom translation tests also use the wasm integration
